+++ plugins/woocommerce-admin/packages/components/src/table/stories/table-summary.js
@@ -2,13 +2,20 @@
  * External dependencies
  */
 import { TableSummary } from '@woocommerce/components';
+import { Card, CardFooter } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import { summary } from './index';
 
-export const Basic = () => <TableSummary data={ summary } />;
+export const Basic = () => (
+	<Card>
+		<CardFooter justify="center">
+			<TableSummary data={ summary } />
+		</CardFooter>
+	</Card>
+);
 
 export default {
 	title: 'WooCommerce Admin/components/TableSummary',