+++ plugins/woocommerce-admin/packages/components/src/table/summary.js
@@ -12,7 +12,7 @@ import PropTypes from 'prop-types';
  */
 const TableSummary = ( { data } ) => {
 	return (
-		<ul className="woocommerce-table__summary">
+		<ul className="woocommerce-table__summary" role="complementary">
 			{ data.map( ( { label, value }, i ) => (
 				<li className="woocommerce-table__summary-item" key={ i }>
 					<span className="woocommerce-table__summary-value">
@@ -35,3 +35,21 @@ TableSummary.propTypes = {
 };
 
 export default TableSummary;
+
+/**
+ * A component to display a placeholder box for `TableSummary`. There is no prop for this component.
+ *
+ * @return {Object} -
+ */
+export const TableSummaryPlaceholder = () => {
+	return (
+		<ul
+			className="woocommerce-table__summary is-loading"
+			role="complementary"
+		>
+			<li className="woocommerce-table__summary-item">
+				<span className="is-placeholder" />
+			</li>
+		</ul>
+	);
+};