+++ plugins/woocommerce-admin/packages/components/src/table/test/index.js
@@ -11,11 +11,12 @@ import userEvent from '@testing-library/user-event';
 import TableCard from '../index';
 import mockHeaders from './data/table-mock-headers';
 import mockData from './data/table-mock-data';
+import mockSummary from './data/table-mock-summary';
 
 expect.extend( { toHaveClass } );
 
 describe( 'TableCard', () => {
-	it( 'should render placeholder table while loading', () => {
+	it( 'should render placeholders for Table and TableSummary while loading', () => {
 		render(
 			<TableCard
 				title="Revenue"
@@ -24,15 +25,22 @@ describe( 'TableCard', () => {
 				rows={ [] }
 				rowsPerPage={ 5 }
 				totalRows={ 5 }
+				summary={ [] }
 			/>
 		);
 
+		// Check Table
 		expect( screen.getByRole( 'group', { hidden: true } ) ).toHaveClass(
+			'is-loading'
+		);
+
+		// Check TableSummary
+		expect( screen.getByRole( 'complementary' ) ).toHaveClass(
 			'is-loading'
 		);
 	} );
 
-	it( 'should not render placeholder table when not loading', () => {
+	it( 'should render table along with summary data when row and summary data is present', () => {
 		render(
 			<TableCard
 				title="Revenue"
@@ -41,10 +49,17 @@ describe( 'TableCard', () => {
 				rows={ mockData }
 				rowsPerPage={ 5 }
 				totalRows={ 5 }
+				summary={ mockSummary }
 			/>
 		);
 
+		// Check Table
 		expect( screen.getByRole( 'group' ) ).not.toHaveClass( 'is-loading' );
+
+		// Check TableSummary
+		expect( screen.getByRole( 'complementary' ) ).not.toHaveClass(
+			'is-loading'
+		);
 	} );
 
 	it( 'should not error with default callback props', () => {
@@ -56,6 +71,7 @@ describe( 'TableCard', () => {
 				rows={ mockData }
 				rowsPerPage={ 1 }
 				totalRows={ 5 }
+				summary={ mockSummary }
 			/>
 		);
 
@@ -85,6 +101,7 @@ describe( 'TableCard', () => {
 				rowsPerPage={ 1 }
 				totalRows={ 5 }
 				rowKey={ ( row ) => row[ 1 ].value }
+				summary={ mockSummary }
 			/>
 		);